Define props for CallRouter, Simulator classes
The call router will need to be able to look through the list of agents,
so it makes sense to have that be a property. The simulator as a whole
consists of a set of consumers, a set of agents, and a call router,
so make those all properties; however, rather than do this in the
`initialize` method, create a separate method that can be called
repeatedly for a given simulator instance.