Add 'satisfied' property to Consumer
A Consumer is either satisfied or unsatisfied. While they are
unsatisfied, they will 'call' the CallRouter [NOT YET IMPLEMENTED]
in order to reach an Agent. Once either
a) the Consumer reaches an Agent on an inbound call, or
b) the Agent calls back the Consumer and reaches them successfully,
the Consumer becomes satisfied.

Define custom reader and accessor methods `satisfied?` and `satisfy`.
Do not add a way to set `@satisfied` to false, since this cannot happen
within the parameters of this simulation.