Give Consumers a 'busy' property
While a Consumer is on the phone with an Agent, they are 'busy'; while
'busy', a Consumer will not make new inbound calls (they also will not
make new inbound calls once they become 'satisfied'). Create
attr_accessor methods for `@busy`, and make an alias method `busy?` in
order to follow best-practice naming conventions for predicate methods.

Additionally, when marking a Consumer as satisfied (via #satisfy), also
set `@busy` to false. While not strictly necessary, it will help ensure
that `busy?` will only return true while the Consumer is actually in a
call with an Agent.