Give Agents a 'busy' property
Create a property for Agents to denote when they are on a call with a
Consumer; when an Agent is 'busy', they cannot accept a new call, so
any new inbound calls must be sent to voicemail.

Because nothing in this simulation needs to be able to change whether or
not an agent is busy, make private setter methods `become_busy` and
`become_free`. While this is more verbose than simply editing `@busy`,
it is better practice to not edit instance variables whenever possible.
Additionally, `become_busy` and `become_free` are a bit more
human-readable.