Add Agent methods for handing incoming calls
Agent#accept_call

Handles what happens when an agent accepts an incoming call from a
Consumer (via the CallRouter) -- namely, the Agent becomes busy, some
time passes (`sleep`), the issue is resolved (i.e. the Consumer becomes
satisfied), and the Agent becomes free again.

Agent#handle_incoming_call

Determines what to do when the Agent is called by the CallRouter; if
that Agent is not busy, accept the call. However, since Ruby is
blocking, we need to create a new thread so that the rest of the code
will not wait for Agent#accent_call to finish its execution.