Define Agent#call_back, #call_back_successfully
Agent#call_back

Represents the Agent calling back the Consumer. Fails 80% of the time;
when it fails, I made the assumption that the Agent would move that
Consumer to the back of the voicemail call-back queue.

Agent#call_back_successfully

A separate method that specifies what happens when the Agent is able to
reach the target Consumer. This method exists in order to facilitate
testing; since Agent#call_back will behave randomly, having a separate
method specifically for the success case obviates a need to try to test
Agent#call_back probabilistically.