Refactor CallRouter specs
Leverage shared examples as much as possible. Rather than manually
construct test cases, use a set of passing and failing values and build
examples of different combinations automatically.

Also, rename the test case context from 'when all agents are available'
to 'when no agents are available', since the method should behave
identically in both contexts, and the latter is more relevant for the
overall behavior of CallRouter.