Add specs for CallRouter#best_agent_for
Specify a method that will look through the sets of free and busy agents
and return the best agent to handle the call (either busy or free).
If any agents are free, this will be the best-matched free agent; if no
agents are free, this will be the best-matched busy agent.

Also, refactor tests for CallRouter#get_best_matched_agent, since the
method will no longer be responsible for dealing with agents being
free/busy.