Give Person a 'running' state
Create a new instance variable, `@is_running`, and use it to start/stop
execution of a new thread that runs indefinitely in the background.
Define a new empty method, Person#act, that the thread can call as long
as it is executing. This method can then be overridden by sub-classes
and replaced with custom behavior specific to the class.