Have Consumer#act call CallRouter#route_call
Allow a CallRouter to be passed in as an argument to Consumer#new, and
update simulation.rb to actually pass the CallRouter in. Then, update
Consumer#act to use this new CallRouter and call #route_call, passing
itself in as an argument.

Until Agent#act is defined, prevent the CallRouter from actually calling
an Agent; instead, just print to the console.