Have Consumers notify Simulator when satisfied
Rather than having the Simulator constantly poll all of the Consumers to
check if they are all satisfied, instead have each Consumer notify the
Simulator once it is satisfied. The Simulator will update its count of
satisfied Consumers, and if all Consumers are satisfied, it will stop
running the simulation.