Have Agents not waste time on satisfied Consumers
If an Agent successfully reaches a Consumer while making a call-back,
if that Consumer has already been satisfied by another Agent,
immediately end the call and move on to the next task. The simulation
parameters do not specify that the Agents can communicate with each
other directly, so this is the only way to prevent duplicate calling.