Define module to export CSV; include in Simulation
Create a module for exporting CSV files and include it in the
Simulator class. Define a generalized 'export' method in the Simulator
class which will call on the module. Create a new directory to hold
exported files and add a `.keep` file; also, create a `.gitignore` and
tell git to ignore the exported files.